:truck: Rename: replaceFragmentWithAnimation 패키지 변경
