+++ app/src/main/java/com/ite/sws/util/replaceFragmentWithAnimation.kt
@@ -1,4 +1,4 @@
-package com.ite.sws.common.data
+package com.ite.sws.util
 
 import androidx.fragment.app.Fragment
